pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "scorevision"
version = "0.1.0"
description = "Score Vision CLI and tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Score Vision" }]

# Core runtime deps
dependencies = [
  "substrate-interface",
  "pydantic>=2",
  "httpx",
  "python-dotenv>=0.21.0",
  "aiohttp>=3.9",
  "Pillow>=10.0",
  "opencv-python>=4.8",
  "click>=8.0.0",
  "huggingface_hub>=0.24",
  "bittensor", 
  "jinja2>=3.1.6",
  "chutes",
  "torch",
  "aiobotocore==2.13.1",
  "numpy>=1.24",
  "pynacl>=1.5",
  "fastapi",
  "uvicorn",
  "petname",
  "ultralytics",
]
 
[project.scripts]
sv = "scorevision:cli"

[project.optional-dependencies]
# Heavy ML stack used by chute templates; optional to keep base lean.
ml = [
  "torch",
  "torchvision",
  "torchaudio",
  "transformers",
]

[tool.hatch.build.targets.wheel]
packages = ["scorevision"]

